requirements.txt
Django==5.1.4
sqlparse==0.5.3
typing_extensions==4.12.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
celery==5.4.0
redis==5.2.1
Pillow==11.1.0
psycopg2==2.9.10
daphne==4.1.2
django-filter==24.3
requests==2.32.3
paho-mqtt== 2.1.0